Extracting the Retrofit client from the Feedback Repository
